refactor(runner): decompose homeboy_refresh god-file test module into submodules
homeboy_refresh.rs was 2623 lines (god_file) with ~1383 lines of
production and a ~1240-line inline test module. Split the test module
into homeboy_refresh/tests/{part_a,part_b}.rs with shared helpers
(ssh_bootstrap_plan, verified_bootstrap_output) promoted to
homeboy_refresh/tests/mod.rs as pub(super).

homeboy_refresh.rs is now 1385 lines (production only). cargo check -p
homeboy-core --lib is clean. No production logic changed. Clears the
god_file finding.
